Migrate Android implementation to MLKit barcode scanner library
The change removes the deprecated Mobile Vision barcode scanning
library [1] and replaces it with the modern MLKit barcode scanner
library [2].

Since the MLKit library harmonizes very well with the input images
produced by the CameraX library [3], the change also replaces the old
SurfaceView based camera view implementation with the modern CameraX
library. As a result the barcode scanning feature will be more robust
on a wider set of devices.

[1] https://developers.google.com/vision/documentation/android/barcodes-overview
[2] https://developers.google.com/ml-kit/vision/barcode-scanning/android
[3] https://developer.android.com/training/camerax